Fixed unused mode hexcodes (#8)
* Fixed unused mode hexcodes

The heat/cool/auto hex codes in the configuration were unused ?!?!? This is to use these to set the mode, and the temp hex to set temperature for units that have seperate codes. 
mode hex codes are optional in the config. Temperature codes will be sent instead assuming these contain the mode information for the AC unit.

* Mode change clean-up

Separated mode change code from sendTemperature to setTargetHeatingCoolingState to allow it to run when only the mode changes.